@@ -1,5 +1,5 @@
 name: qiniu_sdk_base
-version: 0.5.2
+version: 0.6.0
 homepage: https://github.com/qiniu/dart-sdk
 description: The sdk basic of Qiniu products
 
@@ -7,13 +7,13 @@ false_secrets:
   - /test/auth/auth_test.dart
 
 environment:
-  sdk: '>=2.15.0 <4.0.0'
+  sdk: '>=3.0.0 <4.0.0'
 
 dependencies:
   dio: ^5.0.0
   crypto: ^3.0.0
-  meta: ^1.3.0
-  uuid: ^3.0.5
+  meta: ^1.10.0
+  uuid: ^4.4.0
   path: ^1.8.0
 
 dev_dependencies:
