@@ -52,7 +52,7 @@ class UploadPartsTask extends RequestTask<List<Part>> with CacheMixin {
     final keyList = [
       'resource_id/$resourceId',
       'key/$key',
-      'part_size/$partSize'
+      'part_size/$partSize',
     ];
 
     return 'qiniu_dart_sdk_upload_parts_task@[${keyList..join("/")}]';
